service/pyproject.toml
[project]
name = "service"
version = "0.1.0"
description = "划词翻译插件后端 API 服务"
requires-python = ">=3.12.6"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi>=0.133.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
]
